- Some corrections to log output. - Win: TreeViewAdv: added support for handling right clicks on headers. - Win: mforms: implemented handling for header context menu and header title changes.
[TESTS]
  Only tested program flow up to potentially showing a header menu. The only code that uses this feature currently is crashing atm.

+++ library/forms/winforms/src/wf_record_grid_view.cpp
@@ -64,7 +64,7 @@ public:
 
   virtual void set_current_cell(size_t row, int column)
   {
-    _view->set_current_cell(row, column);
+    _view->set_current_cell((int)row, column);
   }
 };
